MainView Cleanup
The MainView class has been modified to better adhere to the Single Responsibility
Principle. In particular, the logic for creating new use case systems (such as those
for the upload, plant view, and plant edit use cases) has been delegated to appropriate
factories. The overlay method in the MainView class has also been changed to behave
in a more expectable way.
